Fix 'in-source build' detection.
This used to check if there was a package manifest in the source
directory, instead of in the build directory. Manifests are expected to be
found in source directories, so that was not correct.

It also made it impossible to quickly build individual packages by the
typical 'mkdir build && cd build && rossum ..' workflow.