Assume 32 bit RG.
Current versions of Roboguide are 32 bit applications and store their
registry keys in the 32 bit view of the registry. Future versions might
become 64 bit applications, which would not require using the 32 bit
view.

For now, assume a 32 bit RG, but allow users to override that by adding
the '--rg64' option to the command line.

If/when RG becomes a 64 bit application, we'll change the logic.